Refactor the processor tests into one file
